requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
boto3==1.34.161
botocore==1.34.161
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
dill==0.3.8
docker==7.1.0
fastapi==0.112.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib-metadata==6.11.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
multiprocess==0.70.16
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pathos==0.3.2
platformdirs==4.2.2
pox==0.3.4
ppft==1.7.6.8
protobuf==4.25.4
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.2
sagemaker==2.228.0
schema==0.7.7
setuptools==72.2.0
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
starlette==0.37.2
tblib==3.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
zipp==3.20.0
